backend/requirements.txt
# backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pydantic==2.4.2
bs4==0.0.1
aiohttp==3.8.5
asyncio==3.4.3
openai==1.12.0
anthropic==0.8.1
google-generativeai==0.3.2
azure-identity==1.15.0
msgraph-sdk==1.0.0
PyPDF2==3.0.1
core_logging